requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
PyMuPDF
python-docx
scikit-learn
